+++ messages/el/base.php
@@ -1,5 +1,6 @@
 <?php
-return array (
+return [
+  '<strong>Translation</strong> Editor' => '',
   '<strong>Translation</strong> History' => '',
   '<strong>Translation</strong> history' => '',
   'Active translation' => '',
@@ -10,10 +11,14 @@ return array (
   'Edit translations' => '',
   'Error while parsing the message, please request support by a translation administrator' => '',
   'File' => 'Αρχείο',
+  'For more informations about translation syntax see' => '',
+  'If the value is empty, the message is considered as not translated.' => '',
   'If you are responsible for this module, try running the following command:' => '',
   'Invalid translation pattern detected, please see {link}' => '',
   'Language' => 'Γλώσσα',
   'Message' => 'Μήνυμα',
+  'Message string can be used with plural forms format. Check i18n section of the documentation for details.' => '',
+  'Messages that no longer need translation will have their translations enclosed between a pair of "@@" marks.' => '',
   'Module' => '',
   'New' => 'Νέα',
   'No translation files found for given selection.' => '',
@@ -22,7 +27,9 @@ return array (
   'Old' => '',
   'Old Translation' => '',
   'Only show missing translations' => '',
+  'Original (en-US)' => '',
   'Otherwise, please report this to the module owner or translation admin.' => '',
+  'Search' => 'Αναζήτηση',
   'Some translations could not be saved.' => '',
   'Some translations may have been purified.' => '',
   'The message file for {settings} not found!' => '',
@@ -33,11 +40,13 @@ return array (
   'The translations for {settings} could not be loaded:' => '',
   'There are unsaved changes, do you really want to leave this page?' => 'Υπάρχουν αλλαγές που δεν έχουν διατηρηθεί, θέλετε σίγουρα να φύγετε από αυτή την σελίδα;',
   'There is no language related space available for language {lang}' => '',
+  'Translated' => '',
   'Translation' => '',
+  'Translation Editor' => '',
   'Translation did not change.' => '',
   'Translations' => '',
   'View History' => '',
   'View history' => '',
   'Your input has been purified.' => '',
   'Your translation seems to be empty and therefore could not be saved.' => '',
-);
+];